+++ app/Http/Controllers/Events/EventCalendar.php
@@ -14,7 +14,7 @@ class EventCalendar extends Controller
         $startOfLastYear = Carbon::now()->subYear()->setDay(1)->setMonth(1)->setTime(0, 0);
 
         $eventList = \App\Models\Event::getAllFiltered(
-            new \App\Models\Filter\EventFilter($startOfLastYear, null, true, false, !auth()->guest())
+            new \App\Models\Filter\EventFilter($startOfLastYear, null, true, false, ! auth()->guest())
         )
             ->get(['title', 'start', 'end', 'whole_day', 'description', 'location', 'dress_code'])
             ->map(function ($event) {